Updated the items to bring issue
